+++ sentry-samples/sentry-samples-spring-boot/src/main/java/io/sentry/samples/spring/boot/PersonService.java
@@ -31,15 +31,10 @@ public class PersonService {
       span.setMeasurement("create_count", createCount);
     }
 
-    Sentry.metrics()
-        .timing(
-            "person.insert",
-            () -> {
-              jdbcTemplate.update(
-                  "insert into person (firstName, lastName) values (?, ?)",
-                  person.getFirstName(),
-                  person.getLastName());
-            });
+    jdbcTemplate.update(
+        "insert into person (firstName, lastName) values (?, ?)",
+        person.getFirstName(),
+        person.getLastName());
     return person;
   }
 }