+++ sentry/src/main/java/io/sentry/HubScopesWrapper.java
@@ -1,6 +1,5 @@
 package io.sentry;
 
-import io.sentry.metrics.MetricsApi;
 import io.sentry.protocol.SentryId;
 import io.sentry.protocol.SentryTransaction;
 import io.sentry.protocol.User;
@@ -337,12 +336,6 @@ public final class HubScopesWrapper implements IHub {
   @Override
   public @Nullable RateLimiter getRateLimiter() {
     return scopes.getRateLimiter();
-  }
-
-  @ApiStatus.Experimental
-  @Override
-  public @NotNull MetricsApi metrics() {
-    return scopes.metrics();
   }
 
   @Override